Добавила модуль констант, убрала лишнее из main.js
